Stop growth by setting blockData.Stage to '1'.
Since the previous version did not clear maxHeight from the
base block (of the plant) it survived if the bamboo
was completly removed and replanted. Changing the state
value of the blockData is more inline how mincraft works.

The topmost block will also be set to have large leaves.

Known issues:
- a placeItem action with bone meal might throw an error
  due to that this event is run async and is not synced
  with the spread evend we're handling.